Adjust with of search drop-downs
By a slight adjustment of the width and padding of the search drop-downs, we avoid that the text input field (to the left) becomes too small in some browser instances.

This is fix though a width adjustment, but it doesn't solve the real issue which is that the CSS is interpreted differently in different browsers, even between individual Chrome installations. An investigation and complete solution is recommended.